feat: 홈 탭에 캐러셀 컴포넌트 추가
