+++ pages/index.tsx
@@ -1,83 +1,36 @@
 import { ampli } from '@/ampli';
 import { useQueryGetGroupBrowsingCard } from '@api/API_LEGACY/meeting/hooks';
-import { useInfinitePosts, useMutationUpdateLike } from '@api/post/hooks';
-import LikeButton from '@components/button/LikeButton';
+import { useInfinitePosts } from '@api/post/hooks';
 import Carousel from '@components/groupBrowsing/Carousel/Carousel';
 import GroupBrowsingSlider from '@components/groupBrowsingSlider/groupBrowsingSlider';
-import FeedItem from '@components/page/meetingDetail/Feed/FeedItem';
-import MeetingInfo from '@components/page/meetingDetail/Feed/FeedItem/MeetingInfo';
 import DesktopFeedListSkeleton from '@components/page/meetingDetail/Feed/Skeleton/DesktopFeedListSkeleton';
 import MobileFeedListSkeleton from '@components/page/meetingDetail/Feed/Skeleton/MobileFeedListSkeleton';
+import RenderPostsWithAds from '@components/page/meetingList/Advertisement';
 import FloatingButton from '@components/page/postList/FloatingButton';
 import { TabList } from '@components/tabList/TabList';
 import { Flex } from '@components/util/layout/Flex';
 import { TAKE_COUNT } from '@constants/feed';
-import { MasonryInfiniteGrid } from '@egjs/react-infinitegrid';
 import { useDisplay } from '@hooks/useDisplay';
 import type { NextPage } from 'next';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 import { useEffect } from 'react';
 import { useInView } from 'react-intersection-observer';
 import { styled } from 'stitches.config';
 
 const Home: NextPage = () => {
-  const { isMobile, isTablet } = useDisplay();
-  const router = useRouter();
+  const { isTablet } = useDisplay();
+
   const { ref, inView } = useInView();
 
-  const { data: postsData, fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } = useInfinitePosts(TAKE_COUNT);
-
-  const { mutate: mutateLikeInAllPost } = useMutationUpdateLike(TAKE_COUNT);
+  const { fetchNextPage, hasNextPage, isFetchingNextPage, isLoading } = useInfinitePosts(TAKE_COUNT);
 
   const { data: groupBrowsingCardData } = useQueryGetGroupBrowsingCard();
-
-  const handleClickLike =
-    (postId: number) => (mutateCb: (postId: number) => void) => (e: React.MouseEvent<HTMLButtonElement>) => {
-      e.preventDefault();
-      ampli.clickFeedlistLike({ location: router.pathname });
-      mutateCb(postId);
-    };
 
   useEffect(() => {
     if (inView && hasNextPage) {
       fetchNextPage();
     }
-  }, [inView, hasNextPage]);
-
-  const renderedPosts = postsData?.pages.map(post => {
-    if (!post) return;
-    return (
-      <Link href={`/post?id=${post?.id}`} key={post?.id}>
-        <FeedItem
-          /* eslint-disable-next-line @typescript-eslint/ban-ts-comment */
-          /* @ts-ignore */
-          post={post}
-          LikeButton={
-            <LikeButton
-              isLiked={post.isLiked}
-              likeCount={post.likeCount}
-              onClickLike={handleClickLike(post.id)(mutateLikeInAllPost)}
-            />
-          }
-          onClick={() =>
-            ampli.clickFeedCard({
-              feed_id: post.id,
-              feed_upload: post.updatedDate,
-              feed_title: post.title,
-              feed_image_total: post.images ? post.images.length : 0,
-              feed_comment_total: post.commentCount,
-              feed_like_total: post.likeCount,
-              group_id: post.meeting.id,
-              platform_type: isMobile ? 'MO' : 'PC',
-              location: router.pathname,
-            })
-          }
-          HeaderSection={<MeetingInfo meetingInfo={post.meeting} />}
-        />
-      </Link>
-    );
-  });
+  }, [inView, hasNextPage, fetchNextPage]);
 
   return (
     <>
@@ -109,7 +62,6 @@ const Home: NextPage = () => {
             </SContentTitle>
             {groupBrowsingCardData && <GroupBrowsingSlider cardList={groupBrowsingCardData}></GroupBrowsingSlider>}
             <SContentTitle style={{ marginBottom: '0px' }}>최신 피드</SContentTitle>
-            <SMobileContainer>{renderedPosts}</SMobileContainer>
           </>
         ) : (
           <>
@@ -131,11 +83,10 @@ const Home: NextPage = () => {
             <Flex align="center" justify="center">
               <SContentTitle style={{ marginBottom: '0px' }}>최신 피드</SContentTitle>
             </Flex>
-            <SDesktopContainer align="center" gap={30}>
-              {renderedPosts}
-            </SDesktopContainer>
           </>
         )}
+
+        <RenderPostsWithAds />
 
         {isFetchingNextPage && isTablet && <MobileFeedListSkeleton count={3} />}
         {!isFetchingNextPage && hasNextPage ? (
@@ -151,31 +102,6 @@ const Home: NextPage = () => {
 };
 
 export default Home;
-
-const SDesktopContainer = styled(MasonryInfiniteGrid, {
-  margin: '$20 0',
-  a: {
-    width: '380px',
-  },
-});
-
-const SMobileContainer = styled('div', {
-  display: 'flex',
-  flexDirection: 'column',
-  margin: 0,
-  '& a:not(:first-child)::before': {
-    content: '',
-    display: 'none',
-
-    '@tablet': {
-      display: 'block',
-      width: '100vw',
-      height: '8px',
-      marginLeft: 'calc(50% - 50vw)',
-      background: '$gray800',
-    },
-  },
-});
 
 const SContentTitle = styled('div', {
   fontStyle: 'H2',
